refactor: move data model to library
