+++ app/package.json
@@ -52,7 +52,6 @@
     "typescript": "4.8.4",
     "web-vitals": "3.0.3",
     "yaml": "2.1.3",
-    "zod": "3.19.1",
     "zustand": "4.1.2"
   },
   "scripts": {
