+++ app/src/features/items/ItemLink.tsx
@@ -2,7 +2,7 @@ import { Descriptions, Popover, Tag, Typography } from 'antd'
 import React from 'react'
 import { Link } from 'react-router-dom'
 import { useStore } from '../../data/store'
-import { ItemId, ItemType } from '../../data/types'
+import { ItemId, ItemType } from '@scope42/data'
 import { getSerialFromId, renderDate } from '../../data/util'
 import { ItemIcon } from '.'
 import { isActive, ItemStatus } from './Status'
