+++ app/src/features/items/editor/DecisionEditor.tsx
@@ -2,7 +2,7 @@ import { Form, Input, message, Modal, Select, Tag } from 'antd'
 import { DecisionIcon } from '../ItemIcon'
 import { Controller, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { DecisionId, DecisionStatus, NewDecision } from '../../../data/types'
+import { DecisionId, DecisionStatus, NewDecision } from '@scope42/data'
 import { DECISION_STATUS_UI } from '../Status'
 import {
   selectAllDecisions,
