+++ app/src/features/items/editor/IssueEditor.tsx
@@ -2,7 +2,7 @@ import { Form, Input, message, Modal, Select, Tag } from 'antd'
 import { IssueIcon } from '../ItemIcon'
 import { Controller, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { IssueId, IssueStatus, NewIssue } from '../../../data/types'
+import { IssueId, IssueStatus, NewIssue } from '@scope42/data'
 import { ISSUE_STATUS_UI } from '../Status'
 import { selectAllIssues, selectAllTags, useStore } from '../../../data/store'
 import { useEditorStore } from './ItemEditor'
