+++ app/src/features/items/table/ItemsTable.tsx
@@ -3,7 +3,7 @@ import { ColumnsType, TableProps } from 'antd/lib/table'
 import dayjs from 'dayjs'
 import React, { useEffect } from 'react'
 import { selectAllTags, useStore } from '../../../data/store'
-import { Item } from '../../../data/types'
+import { Item } from '@scope42/data'
 import { renderDate } from '../../../data/util'
 import { ItemLink } from '../ItemLink'
 import { ItemStatus } from '../Status'
