+++ app/src/features/search/HeaderSearch.tsx
@@ -5,7 +5,7 @@ import { useHotkeys } from 'react-hotkeys-hook'
 import { useNavigate } from 'react-router-dom'
 import { ItemIcon } from '../items'
 import { useStore } from '../../data/store'
-import { ItemId } from '../../data/types'
+import { ItemId } from '@scope42/data'
 import { useNavigateToItem } from '../items'
 import style from './HeaderSearch.module.css'
 import { suggest } from './search-index'
