+++ app/src/pages/decision-details.tsx
@@ -7,7 +7,7 @@ import { DecisionIcon } from '../features/items'
 import { PageHeader } from '../features/layout'
 import { Error404, TicketLink } from '../features/ui'
 import { useParams } from 'react-router-dom'
-import { DecisionId } from '../data/types'
+import { DecisionId } from '@scope42/data'
 import { ItemDetailsPage, useEditorStore } from '../features/items'
 import { Markdown } from '../features/markdown'
 import { Person } from '../features/people'
