+++ app/src/pages/search.tsx
@@ -4,7 +4,7 @@ import { useLocation, useNavigate } from 'react-router-dom'
 import { PageHeader } from '../features/layout'
 import { getAllPossibleStatuses } from '../features/items'
 import { selectAllItems, useStore } from '../data/store'
-import { Item } from '../data/types'
+import { Item } from '@scope42/data'
 import { exists } from '../data/util'
 import { ItemsTablePage } from '../features/items'
 import { search } from '../features/search'
