+++ packages/scope42-data/package.json
@@ -26,5 +26,8 @@
   },
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "zod": "^3.19.1"
   }
 }
