+++ packages/scope42-data/src/index.ts
@@ -1,7 +1,1 @@
-export const add = (a: number, b: number) => {
-  return a + b
-}
-
-export const subtract = (a: number, b: number) => {
-  return a - b
-}
+export * from './model'
