+++ packages/scope42-data/src/model/index.ts
@@ -1,0 +1,16 @@
+import { Decision } from './decision'
+import { Improvement } from './improvement'
+import { Issue } from './issue'
+import { Risk } from './risk'
+
+export * from './workspace-config'
+
+export * from './item-id'
+export { ItemType, Aim42ItemType, Comment } from './item-base'
+
+export * from './issue'
+export * from './risk'
+export * from './improvement'
+export * from './decision'
+
+export type Item = Issue | Risk | Improvement | Decision
