+++ packages/scope42-data/src/model/workspace-config.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod'
+
+export const WorkspaceConfig = z.object({
+  version: z.number().positive().int().default(1)
+})
+
+export type WorkspaceConfig = z.infer<typeof WorkspaceConfig>
