+++ renovate.json
@@ -1,21 +1,16 @@
 {
-  "extends": [
-    "config:base",
-    ":semanticCommitTypeAll(chore)"
-  ],
+  "extends": ["config:base", ":semanticCommitTypeAll(chore)"],
   "stabilityDays": 7,
   "prCreation": "not-pending",
   "dependencyDashboard": true,
   "labels": ["dependencies"],
   "packageRules": [
     {
-      "updateTypes": ["minor", "patch", "pin"],
+      "updateTypes": ["minor", "patch"],
       "automerge": true
     },
     {
-      "packagePatterns": [
-        "^@scope42/"
-      ],
+      "packagePatterns": ["^@scope42/"],
       "stabilityDays": 0
     }
   ]
